Updated support for drawing background drawable when resizing
93bfa3b0bc724a24dd8c4d9f3befba7a32f2ae14 [formerly f3a62fbc58bbc7f081a53248cae48a8951294e8f] added support for drawing
the background draweable when resizing an activity window.
However, after some additional discussion we decided that
R.styleable.Window_windowResizingBackground and
R.integer.config_windowResizingBackgroundColorARGB are not needed.
We use R.styleable.Window_windowBackground for the background
drawable and fallback to using R.styleable.Window_windowBackgroundFallback
which is now public if the first isn't set.

Bug: 24534744
Change-Id: Ia0507e25a1893ea941d259f1d4e88ce500dda154

Former-commit-id: 5bcd3189c52423040281a9f52eeaaef08a7db8a5