Support for activity to opt-in/out of PiP form of multi-window.
While it makes sense to be able to resize most activities types
in multi-window mode. It only makes sense to put specific types
of activities in Picture-in-Picture (PiP) form of multi-window.
For example, activities that play video will be good candidates
while the Settings activity isn't.

The new flag allows the system to differentiate between resizeable
activities that can go into PiP mode and those that can't.

Bug: 25006507
Change-Id: I8ac518cec2fa3c8fb88be40c266b3751fb88f1ce

Former-commit-id: b60692e3d8af750bb089cfc0e0d591002894e118