Show anchored context menu on long press
Also shows anchored menu on D-Pad long press and uses the center of the
view as the anchor point. This is how we already handle hotspot feedback
when there is no explicit center, so there's no visual change there; it's
just more obvious from the View side of things what the result will be.

Bug: 25215353
Change-Id: I930c3aeffc993b7c553ffb626d1b5103c6cb1267

Former-commit-id: 35c6d6b08aa8fc6c9d59c066c9fb80b842f4de01