Add stub isSameTrustConfiguration API
This API will be used in WebView to help determine whether secure
connections to hostname A can be used for secure communication to
hostname B (e.g., HTTP/2 connection pooling).

This is needed because with the new network security configuration a
completely different trust configuration may be used for
foo.com and bar.foo.com, so even if the foo.com certificate contains a
SAN for bar.foo.com it may not be valid for bar.foo.com given the
applications trust configuration.

Change-Id: I87184d392b9a7eca53a9c837996ca7ab5cd5bf12

Former-commit-id: ec95c407b43a03890d3b42b86571110ee90d283d