Implement shell commands for battery and activity services.
The battery service just implements the existing commands that
are available through dump.

The activity service implements the small set of commands that
are available through dump (not the rest of the dump commands),
and also introduces some of the simple "am" shell commands as
a proof-of-concept of moving those into the service implementation.

Change-Id: If5ff80930dde787703e2682e43c36ce1dab05d69

Former-commit-id: 2e44107bf7d9ad97833933422ccd241bf53ff50c