Bandaid over broken keyboard layout selection process.
Automatically select a keyboard layout if we have one that is device
specific and is made for our current locale. Also, provide a way of
requesting layouts for a specific input device rather than just
getting all of them. Custom layouts may not be appropriate for
typical keyboard devices (and custom keyboards may not work with
typical layouts).

Bug: 25062009

Change-Id: I3a0ae5ad68f956b936485791ceb78c347fad7d4f

Former-commit-id: 0748342d2ed264c01384fbaa4446a702a8824813