Fix flicker at the end of docked stack divider animation
Because w.isDragResize() was still true but we already
cleared the flag of the divider controller that it was
resizing, there was a frame at the end of the resize
operation at which all stack surfaces were placed at 0/0,
leading to a flicker.
Fix flicker at the end of docked stack divider animation

Because w.isDragResize() was still true but we already
cleared the flag of the divider controller that it was
resizing, there was a frame at the end of the resize
operation at which all stack surfaces were placed at 0/0,
leading to a flicker.

Change-Id: I126765a31703e5405687bcc511960a41024eb652

Former-commit-id: dcf467ca36130aa719a4954da7578217147b20a0