Set correct focus activity when using FLAG_ACTIVITY_REORDER_TO_FRONT
When FLAG_ACTIVITY_REORDER_TO_FRONT is used to launch an activity we
move the activity to the front of the task, but the ActivityRecord.frontOfTask
isn't set because the activity isn't in the resumed set yet.
This causes issues later on when we try to detemine the focus activity in
ActivityStack.adjustFocusedActivityLocked.
We now set ActivityRecord.frontOfTask for the activity we are moving
due to FLAG_ACTIVITY_REORDER_TO_FRONT

Bug: 25487874
Change-Id: I4a20e6fb11e04ae16361da27c59aec2641b82bfb

Former-commit-id: 6d42df6d176a8defb3e1f1dedc128763a8b91677