Ignore wallpaper when looking for a position for a free window.
Wallpaper wanders through the window list, for example to position
itself directly behind keyguard. Because of this it will break the
ordering based on WindowState.mBaseLayer. There might windows with
higher mBaseLayer behind it and we don't want the new window to
appear above them. An example of this is adding of the docked stack
divider. Consider a scenario with the following ordering (top to
bottom): keyguard, wallpaper, assist preview, apps. We want the dock
divider to land below the assist preview, so the dock divider must
ignore the wallpaper, with which it shares the base layer.

Bug: 25564817
Change-Id: I5bab792d972f845931b54db9f9ae5ff9a4a0e9b3

Former-commit-id: 5e69fda6d9b1a08e2b0c1716ce65e03b2ea1c2d1