Trigger WebView relro creation when new provider package installed.
Previously, the switching of WebView providers was only triggered when a
WebView provider package was replaced. The new channels (e.g. Dev/Beta)
do not exist in the system image which means that we might install a
new WebView provider package without having to replace an existing one
(which before this change meant we didn't trigger the WebView updating
service).

Change-Id: I0933cb5cfc05c64dd950e0b405ede918c2791798

Former-commit-id: 3098cf2ea2d119a5630aa66e617f1f31de38cf17