DO NOT MERGE ANYWHERE
Add ability to set supported audio routes on phone accounts and connection.

The set audio routes are used by Telecom to restrict where the audio may
be routed to. For example, an account can specify that calls may not be
routed over bluetooth headsets, which will prevent a new call from being
routed to this source.

Bug: 25485578
Change-Id: Ia1ca665e9ae8852a2538ab7cbeb3a73217d9b03b

Former-commit-id: dd976b1b0dab6655e180eecd0d2c44658b66f529