Prevent SurfaceView flash while resizing.
We need to make sure we do not preserve
and destroy attached windows when entering
drag resize as we will not swap these windows
to the big surface. On a similar note,
we need to make sure we do not hide child windows
when destroying the parent window for preservation.

Change-Id: I98e02f08d0ea2d5f0032ec7691e865c536df19c6

Former-commit-id: 84b0574cecf7eba313c78fa8d9176ac0d34ef103