Have ControlledInputConnectionWrapper#toString().
'dumpsys input_method' does not generate useful information for
InputMethodManager#mIInputContext.  It is currently just a default
output of Object#toString().

This CL overrides ControlledInputConnectionWrapper#toString() so that
we can easily understand whether this connection is accepting API calls
from the IME or silently droping those API calls.  Note that
ControlledInputConnectionWrapper is not exposed to developers so
overriding #toString() has no effect for application developers.

Bug: 25373872
Change-Id: I89b9d4e15bc1569817bab4d1be429ac6e0462f14

Former-commit-id: 12d66c2cda622fe1df611198d94fdd567adf9fc4