Implement flexible light idle maintenance windows.
For light idle mode, our maintenance window is pretty short (1 minute).
Usually we have nothing to do in it...  but sometimes we may want to
do a sync or download that takes more than a minute, and it would be
nice to not take the overhead of breaking that up in to multiple
windows.

So now we have a flexibile window, from 1 minute to 5 minutes.  We
start out with 1 minute, and any window that uses less than 1
minute increases the available window by that amount for later use.
If we later use more than 1 minute, we correspondingly decrease
the next available window.

Change-Id: Ie273dbd5843b3aa4a3d5d9b2e420cda75517340f

Former-commit-id: 8ed2b97b9e8b8aa34026aa5e9e614494af3d12f6