docs: Fixing malformed hyperlinks
One hyperlink used the {@docRoot} tag even though it's a link to
a separate site (AOSP), so the link didn't work. Two other (internal)
links began with a slash instead of {@docRoot}, which kinda works
but isn't how we do things.

I'll build and test (and post the stage link in my first comment) but
if it looks good I'll go ahead and +2/merge it myself, since there
are no prose changes.

bug: 25755465
Change-Id: I11af855ed7a1e4a12ced7d0d7f6d5a4705c66811

Former-commit-id: e68f6cdabd0ffd5c96935181cfc68b05de0db5ce