Syncronosuly notify SelectPrinterActivity about changes to enabled print services
If the SelectPrinterActivity is visible it shows different emptyState
depending on if there are print services available or not. Hence the
activity has to listen to changes to the list of enabled services.

This also fixes a small syncronsation problem if two observers are
registered for the enabled print services.

Bug: 25666802
Change-Id: I79af66f25f10e66347b48ce9bb99c1657b30a8dd

Former-commit-id: 03755a1f12390be9ddfcce8c5ad3f887fa5ccf9f