Fix tests
The majority of the test failures were around the method assertInstall()
failing a check for a symlink between the data directory in /data/data/
to the application directory in /mnt/asec/. There are a handful of other
failures that will be dealt with separately; those have all been suppressed
from running.

Also took the opportunity of adjusting some of the tests to ensure the
device is properly cleaned up. Packages remained installed after a failure.

Bug: 18048725
Change-Id: Id8f6ac53cfe17de43b5d527993d57812b1a63bef

Former-commit-id: 9f3a78cbaa0e46159dd9925d8acaf57386f05420