Rewrite a fake language code "tl" in SpellChecker.
With following CLs, we already have a special rewrite rule of "tl" to
"fil" for IMEs that are targeting older versions of Android earlier than
Lollipop that did not support three letter language codes and used
"tl" (Tagalog) as the language string for "fil" (Filipino).
  - d3166afaf374ac39e8fe6da2d8bbdcc62c5f69e6 [formerly 92280cd309b0f5967dd253280962d8581844db89] [1]
  - 1faa800bbf713dfcb4e426033c9682c9522d46e0 [formerly ed65bc0c62ca99a118057db7ad54c4ccc14d52d0] [2]

  [1]: I94f203bddceb9c87710cb187cc3cc0ee6d9092a5
  [2]: Ica9cd2baac002c406f92331aadd7725d7424046a

With this CL, we have the same rewrite rule for spell checker services.

Bug: 20696126
Change-Id: I0af0f520a15337e33973391c9965364e3ae1ee4c

Former-commit-id: eae60ba5a4a39d59940e9749a4ad26281d016d03