Don't lose stopping state when getting paused confirmation.
In some cases we request pause and soon after request stop of an
activity. An example of this is maximizing an activity in side by side
mode when recents are visible. First, we add recents to a list of
activities to be stopped (because it becomes invisible), then we request
a pause (because we resume full screen activity and pause all other
stacks) and finally request a stop (from the list to which we added the
request).

The activity now will be put into pausing state (requesting pause), then
stopping (requesting stop), then paused (confirmation comes from the
activity) and stopped (another confirmation comes from activity). If we
switch from stopping to paused, the stop confirmation will become
confused.

Bug: 25729693

Change-Id: I935acd71a28f3d0882f608bdd4d7216cd08ba2eb

Former-commit-id: dba623ae5648923e06157d53fcc0ab8a09a79870