Apply cropping to resizing surfaces.
Apply crop to surfaces even while we resize them. In the case of
SurfaceView, the SurfaceView will be much slower to resize than the
main window. Without crop, this causes the SurfaceView to jut past
the bounds of the main window as it is shrinking. With crop we can
ensure the crop moves with the border in the same SurfaceFlinger
transaction.

Bug: 26010823
Change-Id: Ifb32422de8d18363bd956a457e9efe8cf26678e5

Former-commit-id: 51a1b875c2ba1bf086c849871dd340cfee2ccfaa