Allow direct-callback alarms for non-wakeup alarms
Turns out there are known use cases for it, in particular to avoid
interfering with doze/idle type optimizations.  In addition, it is
not reliably possible to aim for a specific *wall-clock* trigger time
when posting work to a Handler, since the time base used there is
uptime rather than elapsed.

Bug 26114467

Change-Id: I7cede733fc7e09704b63dfe3e98b00365bbdff35

Former-commit-id: ff6390cda43a0333a93c6b87a8eac5484e5ba8e9