Add crop hint and backup-eligible info to wallpaper API
The OS will not distinguish between the physical underlying wallpaper
image and the target displayable subrect of that image.  This will enable
more pleasing conveyance of backed-up wallpaper imagery between dissimilar
devices.  The allow-backup hint is a way for wallpaper sources to ensure
that the OS doesn't back up imagery that should not be propagated across
devices.

(NB: the backing implementation isn't in place yet; this is just to
permit forward-looking work on client apps in the interim.)

Bug 25454501
Bug 25453848

Change-Id: Id014d552ae509f659992d57b915ef95c5d4b8d1a

Former-commit-id: ffa6a886586b581a69e21cf0d831e2e548a25148