Fixed bug with task record not updating to fullscreen sometimes
null is normally passed in to AMS.resizeStack API if the caller wants
to change the stack size to fullscreen. However, it is possible for the
caller to also pass in a bounds that corresponds to the fullscreen bounds
of the display the stack is on. When this happens, the call to window
manager to resize the stack correctly detects that the bounds is equal
to the fullsceen bounds of the display and sets the stack to fullscreen,
but the task record state isn't updated to fullscreen since they were
previously calculated on the activity manager side.
We now check if the bounds corresponds to the fullscreen bounds on the
activity manager side and set it to null so that the task record state
is correctly updated.

Bug: 25683717
Change-Id: Ife753c6e6c034fd8df663ab897d245f1d354bda7

Former-commit-id: b4ec0a312de422440374638195d4709cc74227e9