Fix IMM#showInputMethodAndSubtypeEnabler().
At least on AOSP InputMethodManager#showInputMethodAndSubtypeEnabler()
has been broken since its beginning.  As of Android M, there are three
separate issues:

  1. Type mismatch in message dispatching layer.
  2. It does not correctly reject API calls from clients that do not
     have IME focus.
  3. Context#startActivityAsUser(intent, null, UserHandle.CURRENT) ends
     up with java.lang.SecurityException:
       Permission Denial: startActivity asks to run as user -2 but is
       calling from user 0; this requires
       android.permission.INTERACT_ACROSS_USERS_FULL.

Given that we have a good alternative to achieve the same result, it
would make sense to deprecate that method and encourage developers to
use Intent-based solution instead.

Before doing that, this CL fixes the issues just for the record.

Bug: 26189558
Change-Id: Ic7a0df3203fee19542a3143abba9bc31baf9698a

Former-commit-id: 41f3427381d77fa9fcd12ed820716ebff7011c16