Reset accessibility after Activity enter transition.
Bug 25504648

Invisible views don't respond to accessibility. The transition
system uses a sneaky way to make an invisible view visible
with the GhostView, so the views becoming visible doesn't
trigger the accessibility system to recognize the newly-
visible views. This forces accessibility to rescan the contents
after the transition completes.

Change-Id: If57114b7e7ee304707d53dd235cd7795df8f509f

Former-commit-id: 10d312b5dae0ee565876101b0c914d7a1aab6451