Revert "Camera2: create new streams if surface size has changed"
This change is redundant after the following change is merged:

Camera2: Handle surface size change

When comparing if two OutputConfigurations are equal, also check
the width, height, format, dataspace when the output configurations
were created. With this, CameraDeviceImpl knows that the surface has
changed sizes and will reconfigure the stream.

Bug: 12250682
Change-Id: I4216260a0a08ad87b835ed97e9eaddf1df317fcc
This reverts commit 9a963557617472d9721b3a1af0374537cdfed1e9 [formerly f49f8b0bad0c6cd5db1b603233ad008c5bac4a60].
Former-commit-id: e94b3fe953cc3aa8b3222a85e1004fe769e47877