Create jit profile file during the app startup flow.
ActivityThread has already all the needed information
(userId,sharedAppId) to create the profile file and set the correct
permissions on it. This avoids passing uids to the runtime which now
expects the profile file to be already created.

As part of this change:
- move the registration with the runtim in handleBindApplication
- register only if needed: we could properly setup a profile file and we
have at least one code path. Note that registration with the runtime is
done solely for profiling reasons.

Bug: 26080105
Change-Id: I72657e9ea0a0449a09734456b5f8b5da844cee62

Former-commit-id: 690523969d2f5b202e9929ad9fe0b5cba5537b0f