Remove some minimally supported locales from LocalePicker
For now, these locales are removed:

1. Mundang (mua): the name of the language is all-caps (while the
name of other languages in the locale are either lowercase or
titlecase.

2. Tachelhit (shi): It does not have the name of the Latin or
Tifinagh scripts that it's written in.

3. Vai (vai): It does not have the name of the Latin or Vai scripts
that it's written in.

4. Kashmiri (ks): The spelling of India uses incorrect Unicode
characters, and so do many other pieces of information in the locale.

5. Northen Luri (lrc): It doesn't have the name of the countries it's
spoken in (Iran and Iraq).

(The header of the config file is also fixed.)

Bug: 26279867
Change-Id: I52ad5f9058390f060705e21db62df81513d6ca6e

Former-commit-id: 0a8bf0b36f425dbca46f664c3f5b6af47649f71a