Notify client when the window is moved because of a resize
We need to notify the client that the window has moved if a resize
results in a move without size change. This makes sure that relevent
info on client side (such as mAttachInfo.mWindowLeft/Top) gets updated
to the new frame. Things like View.getLocationOnScreen() may depend
on these to function.

Bug: 25565385
Change-Id: I5b9ded0b16243c14494f9a69257d56570ee8996d

Former-commit-id: bd0d937303ae54d8a5bb5f08080c4164302daefc