docs: Fixing typos
Fixed a reported typo (see bug), and while I had the file open, fixed
a few other typos and formatting problems.

I'll go ahead and build and stage this (see first comment for stage
location) to make sure I didn't break anything, but if it looks okay
I'll +2 it myself and submit it.

bug: 26356168
Change-Id: I57727cecd671ab877c0da27e487fa058d95914c5

Former-commit-id: eab2d05db42418becd0f60499d2b2acd806fea9a