Removed outdated TODO for LocaleList
The suggested TODO is no longer deemed necessary (and is too
expensive to do for all LocaleLists). LocaleLists get (and will get)
passed down to the native level through strings that are the result
of calls to toLanguageTags(), which is pre-computed at construction
time. Users who need to interact with the native level can do their
own caching, such as in commit
08102721779309c73da3aa5429fd536754189c29 [formerly cfc607cf57d453f977c9c4bf09b41ac481c4bbec].

Change-Id: I665f3b28eaf8df309bf1aec96f9d237aeb72865c

Former-commit-id: 8076c3a685c60202a7e11d3fc017ea4922ce6aa2