Implement surfaceRedrawNeeded for GLSurfaceView
GLSurfaceView was not supporting the surfaceRedrawNeeded protocol
which leads to issues with report draw. Thus far this has mostly
been fine as the main surface will not render
a transparent hole until the SurfaceView has finished drawing. However
in the case of SurfaceView replacement the hole in the parent surface
will be preexisting and we must be able to tell that our new SurfaceView
has actually drawn before removing the replacee. This is part of a
series of fixes to 26070641.

Bug: 26070641
Change-Id: Iecc16647f9979aa057449e8bafa540885881ed64

Former-commit-id: 2b3bf720bf28584f700ad8ebb5e8e31a0e466f29