Fix home key behaviour when there are no subdirs in cwd.
The divider is extraneous when a directory contains only subdirs, or
only files.  The "only dirs" case works (the divider is omitted) but in
the case where a directory contains only files, a divider gets inserted
at position 0, which messes up navigation.  This change adds code to
avoid adding a divider when a directory contains only files.

BUG=26406821

Change-Id: I7333911baf94e34bf428a3dd18b55694a73e3a62

Former-commit-id: f5ce1035d41b590c5a25074242baed9dbd48ef3f