Transfer focus on any touch down action
Currently we are deliberately not transfering focus if the ACTION_UP
and ACTION_DOWN points are different. Removing this logic so that we
transfer the focus if user touched within a non-focused task.

If user is scrolling using a pointer device without touch action, it
will not result in focus transfer.

bug: 26040818

Change-Id: If1fd6b8405e1507f60be44f12b6fc599555b8ba7

Former-commit-id: 9817dd098dcb4e37e1bec8b1e44ddb25a89cf8ec