Suppress NPE in IMS#onEvaluateInputViewShown().
This is a follow up CL for the previous commit [1], which may have
triggered an unknown bug in either Android Framework or LatinIME.

  [1]: Id4d332e3909590c68345e10e1f2e18650efb2eb7
       a9093436c30622235611a90f055a34883d731568 [formerly 7b739a802cb7d97460a0bf8a1e9774efe96fa552]

InputMethodService#mSettingsObserver is initialized in #onCreate() and
cleard with null in #onDestroy().  Hence hitting NPE against it implies
that InputMethodService#onEvaluateInputViewShown() can be called before
InputMethodService#onCreate() or after InputMethodService#onDestroy().
Both possibilities are equaly problematic.  Note that this might be a
long-standing issue that just became obvious because of [1].

This CL does not attempt to fix the root cause but just tries to
suppresses the NPE to unblock QA tasks.  A proper fix should be made in
subsequent CLs.

Bug: 22517687
Bug: 26511607
Change-Id: I6bc87c3d18b560fe2253fb9f05557b95b04d0cf0

Former-commit-id: cf8403bc36395eecd9ab03c6b5e128272c3592e9