Change API for killing app-depentents to also kill main app.
In the case where we switch WebView provider we also want to kill the
provider itself in the cases where we want to kill its dependents. This
so that the provider's processes are consistently loading code in the
same way (i.e. either all processes use the webview implementation, or
none of them do).

Bug: 25338573

Change-Id: I2baf7ce1204fb6312a6ad10fa006b789c8ece176

Former-commit-id: 1c7e9ff4a7f11a9f73db0b8e2eb667a049611e49