BT: Don't switch devices when one of the connected BT headsets disconnects
Usecase:
1. Enable multi-hf.
2. Connect to HS1.
3. Connect to HS2.
4. Make a call on AG. Here call audio is present on HS2.
5. Now disconnect HS1 from AG.

Failure:
When HS1 is disconnected, call audio is routed to handset/speaker.

Root cause:
When hs1 disconnection intent is received, SCO path is cleared
and audio is routed to handset/speaker.

Fix:
Check if the device being disconnected is same as the device
call audio is present before clearing SCO path.

Change-Id: If83325679b70b5893e44e8d844000ee028d0246c

Former-commit-id: 2980980a75a0066273683aed8b12a58b212a0487