Stop rebooting into recovery if uncrypt fails.
uncrypt needs to be called when an OTA package sits on /data. If uncrypt
fails or gets skipped (likely due to unexpected runtime restart during
the reboot), we should abort the rebooting. Because otherwise a) it
would fail for sure when reboots into recovery; b) useful logs would be
gone once it reboots.

Bug: 26444951
Change-Id: Ic3bc682b03f003f197e277ed6b6886e85f364754

Former-commit-id: 4fd6c26410cd210a7948f94407d9b2f551badbdc