Voice Messaging Intent API.
  - New activity action for initiating a message to someone by voice.
The message could be text, audio, video or image(photo). The action
supports messaging with a specific contact regardless of the underlying
messaging protocol used.
  - Set of extras to support identification of such contact.

    Primary use case for this action is to be originated by the Voice
Assistant as a voice interaction. In such scenario receiving activity
could verify the source and send a message right away, without any
further input from the user. It allows for a smooth user experience when
sending a message by voice.

Bug: 22791113
Change-Id: I86d55af04d13acbc1390938fa6b914990716cb39

Former-commit-id: 799da391d616685cf011e6a6df34013d9bc536f0