Fix flicker when dismissing non-docked stack
Because the right/bottom side of divider moves offscreen when dismissing non
docked stack, it results in empty bounds which then results in window manager
thinking the just dismissed activity is fullscreen, which leads to a black
flicker. Make sure not to calculate garbage when calculating the bounds for the
non-docked stack.

Bug: 26070457
Change-Id: I65ce9a23dc216971a3ae4df8058d5b86b1b792a5

Former-commit-id: 08b372f71ad24ce29b86ace726bd12d2326ae87e