Deduplicate font file mappings.
With ttc and gx variation fonts, it is now possible and common that a
number of fonts will use the same font file for data but with different
parameters. In the current code each font will map the font file data,
taking up an unecessary amount of virtual address space and is
inefficient with respect to memory management (like the tlb). This CL
deduplicates these file mappings so that a given font file will only be
mapped into memory once.

Change-Id: I5ca69f963a434c72ec4028402ecbf9e0f0ee7148

Former-commit-id: fffcf0a31fd4c9a4ec8aa7de70b1eda0d48fb337