Deliver onEnterAnimationComplete when there is no transition animation.
Documentation says that Activity.onEnterAnimationComplete() will be
called when it's safe to draw, so developers can wait for this callback
to do work. In certain cases we never trigger this callback: when the
screen is frozen and there is no transition animation and the window
just popups into existence. In that case we should forcefully schedule
this callback, so that it can be used reliably.

Bug: 24005761
Change-Id: I2c20ca9bc7fa15ca53d9d492ad1c3f5f9608e8c4

Former-commit-id: 3b81c90efb05c4a3f75cfb39b571dcde4a77c856