Condionally remove the block map file.
We used to unconditionally remove the block map file on boot. Because
the package might be half-way uncrypt'd in a corrupt state. CL in [1]
changes uncrypt to ensure that block.map only gets created at the end of
a successful uncrypt. So we can change to keep the fully uncrypt'd
package and the block map. This is to reduce the work for GmsCore to
avoid re-downloading everything again.

[1]: commit 25dd0386fe69460cd1d39de116197dd2c7bf9ec2.

Bug: 26883096
Change-Id: I58ca22064141bf5d42fa48146a980712c8ce21d9

Former-commit-id: e8217ff4a725004e495ed1506928334f97e5bbf1