Add configurable emergency call number support
In some region, when user dials emergency numbers, the call will go to
IVR system first and then to emergency center. This behavior causes
delay to let consumer to get help from center. Due to region/operator
requirements, some emergency numbers are required to be set-up as non
emeregency numbers on device. On such device, coverting original number
to emergency number is required under the conditions where only
emergency call is allowed - e.g. no SIM. This patch implements this auto
converting logic and its configurations.
E.g. For Taiwan, 110 and 119 need to be converted to 112 under not
in-service state.

Bug: 28123861
Change-Id: I01b3072ae272c1af8d4a19b2fb6ae451b7268821

Former-commit-id: 20c747c038f191f020cc3aacec870ff17370e968